requirements.txt
asgiref==3.8.1
astroid==3.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
dill==0.3.8
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pylint==3.2.5
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tomlkit==0.13.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.6.0
